requirements.txt
alpaca-py==0.13.1
fastapi==0.104.1
numpy==1.24.3
pandas==2.0.3
termcolor==2.3.0
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
transformers==4.31.0
uvicorn==0.24.0.post1
influxdb-client=1.38.0
reactivex=4.0.4
